Handle direct file attachements:
