+++ connectors/package.json
@@ -47,6 +47,7 @@
     "dd-trace": "^5.52.0",
     "eventsource-parser": "^1.0.0",
     "express": "^4.18.2",
+    "file-type": "^20.5.0",
     "fp-ts": "^2.16.0",
     "fs-extra": "^11.1.1",
     "googleapis": "^118.0.0",
